# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Barber Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List approved barber shops
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopListResponse"

    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new barber shop (barber role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopRequest"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/{shopId}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a barber shop by ID
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateShop
      tags: [shops]
      summary: Update a barber shop (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/my:
    get:
      operationId: getMyShop
      tags: [shops]
      summary: Get the current barber's own shop
      responses:
        "200":
          description: My shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopDetail"
        "404":
          description: No shop found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/{shopId}/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

    post:
      operationId: createService
      tags: [services]
      summary: Add a service to a shop (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /shops/{shopId}/services/{serviceId}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
        - name: serviceId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
        - name: serviceId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Shop:
      type: object
      properties:
        id:
          type: number
        ownerId:
          type: number
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - name
        - description
        - address
        - city
        - phone
        - status
        - reviewCount
        - createdAt
    ShopDetail:
      allOf:
        - $ref: "#/components/schemas/Shop"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/Service"
            ownerName:
              type: string
          required:
            - services
            - ownerName
    ShopListResponse:
      type: object
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - shops
        - total
        - page
        - limit
    CreateShopRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
          nullable: true
      required:
        - name
        - description
        - address
        - city
        - phone
    UpdateShopRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
          nullable: true
...
    Service:
      type: object
      properties:
        id:
          type: number
        shopId:
          type: number
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationMinutes:
          type: number
      required:
        - id
        - shopId
        - name
        - price
        - durationMinutes
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationMinutes:
          type: number
      required:
        - name
        - price
        - durationMinutes
...