# lib/api-zod/src/generated/types/createShopRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateShopRequest {
  name: string;
  description: string;
  address: string;
  city: string;
  phone: string;
  imageUrl?: string | null;
}